Move uni_print up to customizations.utils
It is now being used by more than just s3, therefore this moves
it up to the customizations directory.
