+++ .changes/next-release/feature-rds-39317.json
@@ -1,5 +1,5 @@
 {
   "category": "rds",
   "description": "Add command to generate IAM database auth tokens.",
-  "type": "feature"
+  "type": "enhancement"
 }
