+++ awscli/customizations/s3/subcommands.py
@@ -28,9 +28,10 @@ from awscli.customizations.s3.filegenerator import FileGenerator
 from awscli.customizations.s3.fileinfo import FileInfo
 from awscli.customizations.s3.filters import create_filter
 from awscli.customizations.s3.s3handler import S3TransferHandlerFactory
-from awscli.customizations.s3.utils import find_bucket_key, uni_print, \
-    AppendFilter, find_dest_path_comp_key, human_readable_size, \
+from awscli.customizations.s3.utils import find_bucket_key, AppendFilter, \
+    find_dest_path_comp_key, human_readable_size, \
     RequestParamsMapper, split_s3_bucket_key
+from awscli.customizations.utils import uni_print
 from awscli.customizations.s3.syncstrategy.base import MissingFileSync, \
     SizeAndLastModifiedSync, NeverSync
 from awscli.customizations.s3 import transferconfig
