+++ awscli/customizations/s3/utils.py
@@ -16,7 +16,6 @@ from datetime import datetime
 import mimetypes
 import errno
 import os
-import sys
 import time
 from collections import namedtuple, deque
 
@@ -267,49 +266,6 @@ def create_warning(path, error_message, skip_file=True):
     warning_message = WarningResult(message=print_string, error=False,
                                     warning=True)
     return warning_message
-
-
-def uni_print(statement, out_file=None):
-    """
-    This function is used to properly write unicode to a file, usually
-    stdout or stdderr.  It ensures that the proper encoding is used if the
-    statement is not a string type.
-    """
-    if out_file is None:
-        out_file = sys.stdout
-    try:
-        # Otherwise we assume that out_file is a
-        # text writer type that accepts str/unicode instead
-        # of bytes.
-        out_file.write(statement)
-    except UnicodeEncodeError:
-        # Some file like objects like cStringIO will
-        # try to decode as ascii on python2.
-        #
-        # This can also fail if our encoding associated
-        # with the text writer cannot encode the unicode
-        # ``statement`` we've been given.  This commonly
-        # happens on windows where we have some S3 key
-        # previously encoded with utf-8 that can't be
-        # encoded using whatever codepage the user has
-        # configured in their console.
-        #
-        # At this point we've already failed to do what's
-        # been requested.  We now try to make a best effort
-        # attempt at printing the statement to the outfile.
-        # We're using 'ascii' as the default because if the
-        # stream doesn't give us any encoding information
-        # we want to pick an encoding that has the highest
-        # chance of printing successfully.
-        new_encoding = getattr(out_file, 'encoding', 'ascii')
-        # When the output of the aws command is being piped,
-        # ``sys.stdout.encoding`` is ``None``.
-        if new_encoding is None:
-            new_encoding = 'ascii'
-        new_statement = statement.encode(
-            new_encoding, 'replace').decode(new_encoding)
-        out_file.write(new_statement)
-    out_file.flush()
 
 
 class StdoutBytesWriter(object):
